requirements.txt
pymongo
python-dotenv 
voyageai